refacto Cart and Product
