+++ src/main/java/com/Ecommerce/acme/controller/ProductController.java
@@ -20,35 +20,11 @@ public class ProductController {
 	@Autowired
 	private ProductService ps;
 
-	static List<Integer> quantityList = null;
-	static List<Integer> sizeList = null;
-
-	static {
-		quantityList = new ArrayList<>();
-		quantityList.add(1);
-		quantityList.add(10);
-		quantityList.add(50);
-		quantityList.add(100);
-	}
-
-	static {
-		sizeList = new ArrayList<>();
-		sizeList.add(41);
-		sizeList.add(42);
-		sizeList.add(43);
-		sizeList.add(44);
-		sizeList.add(45);
-		sizeList.add(46);
-	}
-
 	@GetMapping("/products")
-	public String getHomePage(Model model) {
-		model.addAttribute("products", ps.getAllProduct());
-		model.addAttribute("quantityList", quantityList);
-		model.addAttribute("sizeList", sizeList);
-
-		return "products";
+	public String showProducts(Model model) {
+		return ps.getListProduct(model);
 	}
+	
 
 	@PostMapping("/products")
 	public String addSelectToCart(Selection selection, User user, Product product, Authentication authentication, Model model, BindingResult bindingResult) {
